SUPERNOVA: Move part 1 files to it's own directory
