+++ engines/supernova/supernova.cpp
@@ -45,7 +45,7 @@
 #include "supernova/screen.h"
 #include "supernova/sound.h"
 #include "supernova/supernova.h"
-#include "supernova/state.h"
+#include "supernova/supernova1/state.h"
 #include "supernova/game-manager.h"
 
 namespace Supernova {
