+++ engines/supernova/supernova1/rooms.h
